[refactor]Remove unused variables and refactor common configurations
